requirements.txt
Django==2.1.5
psycopg2-binary
pandas
djangorestframework